Split Apex chart options in separate files for clarity
