+++ front/src/components/boxs/chart/ApexChartAreaOptions.js
@@ -1,0 +1,64 @@
+const getApexChartAreaOptions = ({ displayAxes, height, series, COLORS }) => {
+  const options = {
+    chart: {
+      type: 'area',
+      fontFamily: 'inherit',
+      height,
+      parentHeightOffset: 0,
+      sparkline: {
+        enabled: !displayAxes
+      },
+      toolbar: {
+        show: false
+      },
+      animations: {
+        enabled: false
+      }
+    },
+    dataLabels: {
+      enabled: false
+    },
+    fill: {
+      opacity: 0.16,
+      type: 'solid'
+    },
+    stroke: {
+      width: 2,
+      lineCap: 'round',
+      curve: 'smooth'
+    },
+    series,
+    grid: {
+      strokeDashArray: 4,
+      padding: {
+        left: -4
+      }
+    },
+    xaxis: {
+      labels: {
+        padding: 0,
+        datetimeUTC: false
+      },
+      tooltip: {
+        enabled: false
+      },
+      axisBorder: {
+        show: false
+      },
+      type: 'datetime'
+    },
+    yaxis: {
+      labels: {
+        padding: 4
+      }
+    },
+    colors: COLORS,
+    legend: {
+      show: displayAxes,
+      position: 'bottom'
+    }
+  };
+  return options;
+};
+
+export { getApexChartAreaOptions };
