+++ front/src/components/boxs/chart/ApexChartStepLineOptions.js
@@ -1,0 +1,62 @@
+const getApexChartStepLineOptions = ({ height, displayAxes, series, COLORS }) => {
+  const options = {
+    chart: {
+      type: 'line',
+      fontFamily: 'inherit',
+      height,
+      parentHeightOffset: 0,
+      sparkline: {
+        enabled: !displayAxes
+      },
+      toolbar: {
+        show: false
+      },
+      animations: {
+        enabled: false
+      }
+    },
+    dataLabels: {
+      enabled: false
+    },
+    fill: {
+      opacity: 1
+    },
+    stroke: {
+      width: 2,
+      curve: 'stepline'
+    },
+    series,
+    grid: {
+      strokeDashArray: 4,
+      padding: {
+        left: -4
+      }
+    },
+    xaxis: {
+      labels: {
+        padding: 0,
+        datetimeUTC: false
+      },
+      tooltip: {
+        enabled: false
+      },
+      axisBorder: {
+        show: false
+      },
+      type: 'datetime'
+    },
+    yaxis: {
+      labels: {
+        padding: 4
+      }
+    },
+    colors: COLORS,
+    legend: {
+      show: displayAxes,
+      position: 'bottom'
+    }
+  };
+  return options;
+};
+
+export { getApexChartStepLineOptions };
